Module serde_url_params::ser

Serialize a Rust data structure into URL parameters string.

Structs

Serializer
A structure for serializing Rust values into URL parameters string.

Functions

to_string
Serialize the given data structure as a String of URL parameters.

to_vec
Serialize the given data structure as a byte vector containing URL parameters.

to_writer
Serialize the given data structure as URL parameters into the IO stream.